+++ scripts/com.r2studio.TsumBeta/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the TsumBeta script will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [v74] - tba
+
+### Fixed
+- JP Tsum Tsum update broke "Receive All Hearts". As there are now more features, "Receive All Hearts" now respects the "Skip Ruby" option. 
+
+
 ## [v73] - 2024-12-19
 
 ### Fixed